requirements.txt
aiohttp>=3.9.3
ntgcalls>=1.2.0
pyrogram>=1.2.20
motor==3.0.0
TgCrypto
pytz
# uvloop
redis
redis[hiredis]
python-dotenv
pytimeparse
pyromod==1.1
# unidecode
python-telegram-bot
bs4
python-dateutil
# requests
# uvloop
# deprecation
screeninfo
# Pillow
# psutil
# pytgcalls==3.0.0.dev24
# git+https://github.com/XhaidarX00/pytgcall
# googletrans==4.0.0-rc1
# emoji
# pytube
# youtube-search-python
# pydub
# SpeechRecognition
# nltk
# Sastrawi
# gtts
# google-generativeai
# ffmpeg-python
# yt-dlp
# SpeechRecognition